Refactor some rawdb methods
